# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Torneo Solidario",
        "description": "API del torneo solidario",
        "contact": {
            "email": "minillanillo@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/equipos": {
            "get": {
                "tags": [
                    "equipos"
                ],
                "summary": "Obtener todos los equipos de la web",
                "description": "Obtener todos los equipos en la llamada a la API",
                "operationId": "indexEquipos",
                "responses": {
                    "200": {
                        "description": "Equipo disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipos disponibles"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Equipo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay equipos encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No hay equipos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "equipos"
                ],
                "summary": "Crear un equipo con sus jugadores",
                "description": "Crear un equipo con sus jugadores",
                "operationId": "storeEquipo",
                "requestBody": {
                    "description": "Datos del equipo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "jugadores"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Equipo 1"
                                    },
                                    "centro_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "jugadores": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Jugador"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipo creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo creado correctamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Equipo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/api/equipos/{id}": {
            "get": {
                "tags": [
                    "equipos"
                ],
                "summary": "Obtener un equipo",
                "description": "Obtener un equipo por su id",
                "operationId": "showEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo encontrado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Equipo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "equipos"
                ],
                "summary": "Actualizar un equipo",
                "description": "Actualizar un equipo",
                "operationId": "updateEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del equipo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Equipo 1"
                                    },
                                    "centro_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipo actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo actualizado correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Equipo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "equipos"
                ],
                "summary": "Eliminar un equipo",
                "description": "Eliminar un equipo por su id",
                "operationId": "deleteEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipo eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Centro": {
                "title": "Centro",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "IES Ejemplo"
                    }
                },
                "type": "object"
            },
            "Ciclo": {
                "title": "Ciclo",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "FPGB Tapicería y Cortinaje"
                    },
                    "familia": {
                        "$ref": "#/components/schemas/Familia"
                    }
                },
                "type": "object"
            },
            "Equipo": {
                "title": "Equipo",
                "required": [
                    "nombre",
                    "jugadores"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Desguace FC"
                    },
                    "grupo": {
                        "type": "string",
                        "example": "A"
                    },
                    "centro": {
                        "$ref": "#/components/schemas/Centro"
                    },
                    "jugadores": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Jugador"
                        }
                    }
                },
                "type": "object"
            },
            "Estudio": {
                "title": "Estudio",
                "properties": {
                    "centro": {
                        "type": "string",
                        "example": "IES Ejemplo"
                    },
                    "curso": {
                        "type": "integer",
                        "example": 1
                    },
                    "ciclos": {
                        "$ref": "#/components/schemas/Ciclo"
                    }
                },
                "type": "object"
            },
            "Familia": {
                "title": "Familia",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Textil, Confección y Piel"
                    }
                },
                "type": "object"
            },
            "Jugador": {
                "title": "Jugador",
                "required": [
                    "nombre",
                    "equipo_id"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Álvaro"
                    },
                    "apellido1": {
                        "type": "string",
                        "example": "Ruiz"
                    },
                    "apellido2": {
                        "type": "string",
                        "example": "Gutierrez"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "jugador/capitan/entrenador"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678A"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@exa.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+34 666 666 666"
                    },
                    "goles": {
                        "type": "intenger",
                        "example": 1
                    },
                    "asistencias": {
                        "type": "integer",
                        "example": 1
                    },
                    "tarjetas_amarillas": {
                        "type": "integer",
                        "example": 1
                    },
                    "tarjetas_rojas": {
                        "type": "integer",
                        "example": 1
                    },
                    "lesiones": {
                        "type": "integer",
                        "example": 1
                    },
                    "equipo_nombre": {
                        "type": "string",
                        "example": "Equipo Example"
                    },
                    "estudios": {
                        "$ref": "#/components/schemas/Estudio"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to authenticate",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}